requirements.txt
pillow==4.0.0